XXX Task logging improved

+++ model/tasks/delete-temporal-files-task.js
@@ -1,22 +1,17 @@
-var name = 'task:deleteTemporalFiles';
-var PRESERVE_TIME  = 1000 * 60 * 60 * 48;  //48 horas
-//
-var utils = require('../utils');
-var _ = require('lodash');
-var moment = require('moment');
-var ctrl = require('../db.controller').create;
-var fs = require('fs');
 var path = require('path');
-var log = (m) => {
-    console.log(name + ':' + m);
-}
-var logSave = (msg, type) => ctrl('Log').save({
-    message: msg,
-    type: type
+var resolver = require(path.join(process.cwd(), 'model/facades/resolver-facade'));
+var utils = require('../utils');
+var fs = require('fs');
+
+const PRESERVE_TIME = 1000 * 60 * 60 * 48; //48 horas
+
+var logger = resolver.loggerFacade({
+    name: "TASKS",
+    category: 'CLEANING'
 });
-var ensureDirectory = (path) => {
-    if (!fs.existsSync(path))
-        fs.mkdirSync(path);
+
+function debugTerminal() {
+    logger.debugTerminal.apply(logger, Array.prototype.slice.call(arguments));
 }
 
 function deleteFile(path) {
@@ -25,17 +20,16 @@ function deleteFile(path) {
             fs.unlink(path);
         }
     }
-    catch (e) {};
+    catch (e) {}
 }
 
 function replaceAll(word, search, replacement) {
     return word.replace(new RegExp(search, 'g'), replacement);
-};
+}
 
-var _testDone = true;
+
 
 module.exports = {
-    name: name,
     interval: 1000 * 60 * 60, //each hour
     handler: handler,
     runAtStartup: true
@@ -43,32 +37,30 @@ module.exports = {
 
 
 function handler(data, cb) {
-    log('start');
-
-    //ensureDirectory(process.cwd() + '/www');
-    //ensureDirectory(process.cwd() + '/www/temp');
     var _path = utils.getFileTempPath();
-
-    if (!_testDone) {
-        fs.writeFileSync(path.join(_path, "test_" + (new Date().getTime() - (1000 * 60)).toString() + '.pdf'), '');
-        _testDone = true;
-    }
-
-
+    
+    debugTerminal('Reading',_path);
+    
 
     fs.readdir(_path, function(err, files) {
         if (err) {
-            throw err;
+            return logger.errorTerminal(err);
         }
-        files.map(function(file) {
+        files = files.map(function(file) {
             //return path.join(_path, file);
             return file;
         }).filter(function(file) {
             return fs.statSync(path.join(_path, file)).isFile();
-        }).forEach(function(file) {
+        })
+        
+        if(files.length===0){
+            return logger.debugTerminal('No temporal files present.');
+        }
+
+        files.forEach(function(file) {
             //log('reading:ext:'+path.extname(file));
             if (path.extname(file) !== '.pdf') {
-                log('file:deleting[not-a-pdf-file]');
+                debugTerminal('deleting[not-a-pdf-file]');
                 deleteFile(path.join(_path, file));
             }
             else {
@@ -79,25 +71,24 @@ function handler(data, cb) {
                     var d = new Date(parseInt(str));
                     if (isFinite(d)) {
                         if (Date.now() - d > PRESERVE_TIME) {
-                            log('file:deleting[createdAt > '+(Date.now() - d) / 1000 + ']');
+                            debugTerminal('deleting[createdAt > ' + (Date.now() - d) / 1000 + ']');
                             deleteFile(path.join(_path, file));
                         }
                         else {
-                            log('file:waiting[createdAt ' + (Date.now() - d) / 1000 + '-secs-ago][to-be-deleted-in '+Math.abs(((Date.now() - d) - PRESERVE_TIME)/1000)+'-secs]');
+                            debugTerminal('waiting[createdAt ' + (Date.now() - d) / 1000 + '-secs-ago][to-be-deleted-in ' + Math.abs(((Date.now() - d) - PRESERVE_TIME) / 1000) + '-secs]');
                         }
                     }
                     else {
-                        log('file:deleting[not-a-valid-date]');
+                        debugTerminal('deleting[not-a-valid-date]');
                         deleteFile(path.join(_path, file));
                     }
                 }
                 catch (e) {
                     console.log(e);
-                    log('file:deleting[exception]');
+                    debugTerminal('deleting[exception]');
                     deleteFile(path.join(_path, file));
                 }
             }
         });
     });
 }
-
